Add files via upload
This program uses a class data structure to traverse through the ordered list. I then implemented many functions including print_list to print the list of numbers added to the list, duplicate function to count the amount of duplicates in the list and return them, a pop_pos function which takes a number as an input to remove and print the index number inputted from the list, an index function that take an input and return the integer. I then initialized a list called aList added a list of integers to it then called each function to test them.